install: symlink local scripts/binaries into ~/.bin
Move bin/ directory into .bin/ to "keep it out of the way".

Another option is to move it to ~/.local/bin but there is no guarantee that
this directory wouldn't be used in the future for XDG things. So let's
keep things separate for the timebeing.

See https://lists.fedoraproject.org/pipermail/devel/2011-July/154855.html for
some historic discussion.